+++ modules/cpu/texture/Texture2DShared.h
@@ -6,6 +6,7 @@
 #include "TextureShared.h"
 #include "ospray/OSPEnums.h"
 
+#define MAX_MIPMAP_LEVEL 32
 #ifdef __cplusplus
 namespace ispc {
 #endif // __cplusplus
@@ -15,10 +16,10 @@ struct Texture2D
   Texture super;
 
   vec2i size;
-  vec2f sizef; // size, as floats; slightly smaller than 'size' to avoid range
-               // checks
-  vec2f halfTexel; // 0.5/size, needed for bilinear filtering and clamp-to-edge
-  void *data;
+
+  void *data[MAX_MIPMAP_LEVEL];
+  int maxLevel;
+  vec4f avg; // used instead of last level
 
   OSPTextureFormat format;
   OSPTextureFilter filter;
@@ -27,9 +28,9 @@ struct Texture2D
 #ifdef __cplusplus
   Texture2D()
       : size(0),
-        sizef(0.f),
-        halfTexel(0.f),
-        data(nullptr),
+        data{nullptr},
+        maxLevel(0),
+        avg(0.f, 0.f, 0.f, 1.f),
         format(OSP_TEXTURE_FORMAT_INVALID),
         filter(OSP_TEXTURE_FILTER_LINEAR),
         wrapMode(vec2ui(OSP_TEXTURE_WRAP_REPEAT))
@@ -37,10 +38,12 @@ struct Texture2D
     super.type = TEXTURE_TYPE_2D;
   }
   void set(const vec2i &aSize,
-      void *aData,
+      void **aData,
+      int maxLevel,
       OSPTextureFormat format,
       OSPTextureFilter flags,
-      vec2ui wrapMode = vec2ui(OSP_TEXTURE_WRAP_REPEAT));
+      const vec2ui &wrapMode,
+      const vec4f &avg = {0.f, 0.f, 0.f, 1.f});
 };
 } // namespace ispc
 #else